Bug 845080 - Extract backgroundDataIsEnabled() from AnnouncementsService to BackgroundService. r=rnewman
